pyproject.toml
[project]
name = "new-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "pywin32>=311",
]
